[S07P31D201-114] UPDATE : 민원 | 안전신문고 페이지 이동

+++ AOS/EyesOn/data/src/main/java/com/d201/data/api/ComplaintsApi.kt
@@ -21,7 +21,7 @@ interface ComplaintsApi {
     suspend fun selectComplaintsByBlind(@Query("page") page: Int, @Query("size") size: Int): BaseResponse<PagingResult<ComplaintsResponse>>
 
     @GET("complaints/{complaintsSeq}")
-    suspend fun selectComplaintsBySeq(@Path("complaintsSeq") seq: Int): BaseResponse<ComplaintsResponse>
+    suspend fun selectComplaintsBySeq(@Path("complaintsSeq") seq: Long): BaseResponse<ComplaintsResponse>
 
     @PUT("complaints/return")
     suspend fun returnComplaints(@Body complaintsRequest: ComplaintsRequest): BaseResponse<Void>
