+++ AOS/EyesOn/domain/src/main/java/com/d201/domain/repository/ComplaintsRepository.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.flow.Flow
 interface ComplaintsRepository {
     fun insertComp(complaintsRequest: Complaints): Flow<ResultType<BaseResponse<Void>>>
 
-    fun selectComplaintsBySeq(seq: Int): Flow<ResultType<BaseResponse<Complaints>>>
+    fun selectComplaintsBySeq(seq: Long): Flow<ResultType<BaseResponse<Complaints>>>
 
     fun returnComplaints(complaintsRequest: Complaints): Flow<ResultType<BaseResponse<Void>>>
